capstone-projects/vuln-intel-platform/src/main.py
#!/usr/bin/env python3
"""
Vulnerability Intelligence Platform - Main Entry Point

FastAPI server for vulnerability management with AI-powered analysis.
"""

from dotenv import load_dotenv
from rich.console import Console

load_dotenv()
console = Console()

try:
    from fastapi import FastAPI, HTTPException
    from pydantic import BaseModel
    import uvicorn
    FASTAPI_AVAILABLE = True
except ImportError:
    FASTAPI_AVAILABLE = False
    console.print("[yellow]FastAPI not installed. Run: pip install fastapi uvicorn[/yellow]")


if FASTAPI_AVAILABLE:
    app = FastAPI(
        title="Vulnerability Intelligence Platform",
        description="AI-powered vulnerability management and prioritization",
        version="0.1.0"
    )

    # Pydantic models
    class CVEQuery(BaseModel):
        query: str

    class Asset(BaseModel):
        hostname: str
        ip: str
        os: str
        criticality: str = "medium"

    # API Endpoints
    @app.get("/")
    async def root():
        return {
            "name": "Vulnerability Intelligence Platform",
            "version": "0.1.0",
            "status": "running"
        }

    @app.get("/api/cves")
    async def list_cves(limit: int = 10, severity: str = None):
        """List recent CVEs."""
        # TODO: Implement CVE data retrieval
        return {
            "message": "TODO: Implement CVE listing",
            "limit": limit,
            "severity": severity
        }

    @app.get("/api/cves/{cve_id}")
    async def get_cve(cve_id: str):
        """Get CVE details."""
        # TODO: Implement CVE lookup
        return {
            "cve_id": cve_id,
            "message": "TODO: Implement CVE details"
        }

    @app.post("/api/assets")
    async def add_asset(asset: Asset):
        """Add an asset to inventory."""
        # TODO: Implement asset management
        return {
            "message": "TODO: Implement asset management",
            "asset": asset.model_dump()
        }

    @app.post("/api/query")
    async def query_cves(query: CVEQuery):
        """RAG-powered CVE query."""
        # TODO: Implement RAG query
        return {
            "query": query.query,
            "message": "TODO: Implement RAG-powered Q&A"
        }

    @app.get("/api/reports/executive")
    async def executive_report():
        """Generate executive vulnerability report."""
        # TODO: Implement report generation
        return {
            "message": "TODO: Implement executive report generation"
        }


def main():
    """Run the API server."""
    console.print("[bold cyan]Vulnerability Intelligence Platform[/bold cyan]")
    console.print("Starting API server...")

    if FASTAPI_AVAILABLE:
        console.print("\nAPI endpoints:")
        console.print("  GET  /api/cves          - List CVEs")
        console.print("  GET  /api/cves/{id}     - Get CVE details")
        console.print("  POST /api/assets        - Add asset")
        console.print("  POST /api/query         - RAG query")
        console.print("  GET  /api/reports/exec  - Executive report")
        console.print("\nStarting server on http://localhost:8000")
        console.print("API docs at http://localhost:8000/docs\n")

        uvicorn.run(app, host="0.0.0.0", port=8000)
    else:
        console.print("[red]Cannot start server without FastAPI[/red]")


if __name__ == "__main__":
    main()
