capstone-projects/vuln-intel-platform/requirements.txt
# Core
langchain>=0.1.0
langchain-anthropic>=0.1.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
chromadb>=0.4.0

# Data
pandas>=2.0.0
requests>=2.31.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Utils
python-dotenv>=1.0.0
rich>=13.0.0

# Testing
pytest>=7.4.0
httpx>=0.26.0
